actualizacion

+++ BD/carrera_curso_monto.sql
@@ -1,1 +1,4 @@
-ALTER TABLE `p_horarios_programados` ADD `carrera` INT NULL AFTER `tolerancia`, ADD `curso` INT NULL AFTER `carrera`, ADD `monto_hora` INT NULL AFTER `curso`;  
+ALTER TABLE `p_horarios_programados` 
+ADD `carrera` INT NULL AFTER `tolerancia`, 
+ADD `curso` INT NULL AFTER `carrera`, 
+ADD `monto_hora` decimal(10,2) NULL AFTER `curso`;
